+++ services/api/src/controllers/health_check.rs
@@ -2,7 +2,7 @@ use crate::{
     utils::{
         generate_id::{Entities, PlutomiId},
         get_current_time::iso_format,
-        logger::{self, LogLevel, LogObject},
+        logger::{LogLevel, LogObject},
     },
     AppState,
 };
