--- services/api/src/entities.rs
@@ -1,5 +1,0 @@
-pub mod user;
-
-pub enum EntityType {
-    User(user::User),
-}
