+++ services/events-consumer/src/main.rs
@@ -1,3 +1,11 @@
 fn main() {
     println!("Hello, world!");
+
+    // TODO: Acceptance criteria
+    // - [ ] Can read .env parsed
+    // - [ ] Can connect to MongoDB
+    // - [ ] Can log messages
+    // - [ ] Can consume events / consumer groups
+    // - [ ] Can process/ publish  events
+    // - [ ] Can generate ids with shared lib
 }
