+++ services/shared/Cargo.toml
@@ -1,10 +1,8 @@
 [package]
-name = "events-consumer"
+name = "shared"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-# Shared utils
-shared = { path = "../shared" }
